requirements.txt
google-generativeai
pydantic
python-dotenv
colorama
rich
